+++ frontend/app/src/components/home/Markdown.svelte
@@ -17,6 +17,7 @@
 
     $: singleEmoji = isSingleEmoji(text);
     $: userStore = client.userStore;
+    $: communityMembers = client.currentCommunityMembers;
     $: options = {
         breaks: !oneLine,
     };
@@ -25,7 +26,7 @@
         return text.replace(/@UserId\(([\d\w-]+)\)/g, (match, p1) => {
             const u = $userStore[p1];
             if (u !== undefined) {
-                const displayName = client.getDisplayName(u);
+                const displayName = client.getDisplayName(u, $communityMembers);
                 return `**[@${displayName}](/user/${u.userId})**`;
             }
             return match;